backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
python-dotenv==1.0.0
PyMySQL==1.1.0
bcrypt==4.1.3
bleach==6.1.0